buy/stake tests are working !! 10/10
