+++ test/bankTest.ts
@@ -41,12 +41,13 @@ contract('Bank', async (accounts: string[]) => {
     const USDC_FIX_6MTH_CLASS_ID = 1;
     const USDT_FIX_6MTH_CLASS_ID = 2;
 
-    it('Initialisation', async () => {
+    before('Initialisation', async () => {
         usdcContract = await USDC.deployed();
         usdtContract = await USDT.deployed();
         bankContract = await Bank.deployed();
         dbitContract = await DBIT.deployed();
         wethContract = await WETH.deployed();
+        dgovContract = await DGOV.deployed();
         apmContract = await APM.deployed();
         bondContract = await DebondBondTest.deployed();
     })
@@ -55,46 +56,11 @@ contract('Bank', async (accounts: string[]) => {
         const classes = (await bankContract.getClasses()).map(c => c.toNumber())
         console.log("classes: " + classes)
     })
-
-    it('buy Bonds with USDC', async () => {
-
-
-        await usdcContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
-        await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        await bankContract.buyBond(USDC_FIX_6MTH_CLASS_ID, DBIT_FIX_6MTH_CLASS_ID, web3.utils.toWei('3000', 'ether'), PurchaseMethod.BUYING, 0, {from: buyer});
-        const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
-        console.log("balance Bond D/BIT: AFTER " + (await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0])));
-
-        console.log("1");
-        await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        console.log("2");
-        await bankContract.buyBond(USDC_FIX_6MTH_CLASS_ID, DBIT_FIX_6MTH_CLASS_ID, web3.utils.toWei('3000', 'ether'), PurchaseMethod.BUYING,20, {from: buyer});
-        console.log("3");
-
-
-    })
-
-    it('buy Bonds with USDT', async () => {
-
-        usdcContract = await USDC.deployed();
-        usdtContract = await USDT.deployed();
-        bankContract = await Bank.deployed();
-        dbitContract = await DBIT.deployed();
-        apmContract = await APM.deployed();
-        bondContract = await DebondBondTest.deployed();
-
-        await usdtContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
-        await usdtContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        await bankContract.buyBond(USDT_FIX_6MTH_CLASS_ID, DBIT_FIX_6MTH_CLASS_ID, web3.utils.toWei('3000', 'ether'), PurchaseMethod.BUYING,0, {from: buyer});
-        const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
-        console.log("balance Bond D/BIT: AFTER " + (await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0])));
-
-
-
-    })
+    
     //todo : put comments for test
-    //we should do this test with something else than usdc
-    it('buy Bonds: stakeForDbitBondWithElse', async () => {
+    
+
+    it('stakeForDbitBondWithElse', async () => {
 
         //mint usdc to the buyer
         await usdcContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
@@ -105,7 +71,7 @@ contract('Bank', async (accounts: string[]) => {
 
         //approve usdc and buy bonds
         await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        let amount = await bankContract.stakeForDbitBondWithElse(USDC_FIX_6MTH_CLASS_ID, DBIT_FIX_6MTH_CLASS_ID, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        await bankContract.stakeForDbitBondWithElse(1, 0, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
         
        //log his nonce so we can use it to query bond blance
         const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
@@ -128,8 +94,12 @@ contract('Bank', async (accounts: string[]) => {
         //we should have 3000 usdc bond
         expect( UsdcbondBalance.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
         
+        //we should have more than 3% bond
+        expect( parseFloat(web3.utils.fromWei(dbitbondBalance, "ether"))).to.greaterThan(90);
+
         //now we have check bond balances, we check balances in apm
 
+        
         const s = await apmContract.getReserves(usdcContract.address, dbitContract.address);
         console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
 
@@ -139,22 +109,15 @@ contract('Bank', async (accounts: string[]) => {
 
         let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
         console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+        
     })
 
-    it.only('buy Bonds: stakeForDgovBondWithDbit', async () => {
+    it('stakeForDgovBondWithDbit', async () => {
 
-        usdcContract = await USDC.deployed();
-        usdtContract = await USDT.deployed();
-        bankContract = await Bank.deployed();
-        dbitContract = await DBIT.deployed();
-        wethContract = await WETH.deployed();
-        apmContract = await APM.deployed();
-        bondContract = await DebondBondTest.deployed();
-
-
-        await dbitContract.setBankAddress(bankContract.address, {from: accounts[0]});
+        await dbitContract.setBankAddress(accounts[1]);
         //mint dbit to the buyer
-        await dbitContract.mintCollateralisedSupply(buyer, web3.utils.toWei('100000', 'ether'), {from : bankContract.address});
+        await dbitContract.mintCollateralisedSupply(buyer, web3.utils.toWei('100000', 'ether'), {from: accounts[1]});
+        await dbitContract.setBankAddress(bankContract.address)
 
         //log his dbit balance
         let balance = await dbitContract.balanceOf(buyer);
@@ -162,10 +125,10 @@ contract('Bank', async (accounts: string[]) => {
 
         //approve dbit and buy bonds
         await dbitContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        let amount = await bankContract.stakeForDgovBondWithDbit(0, 4, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        await bankContract.stakeForDgovBondWithDbit(0, 4, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
         
        //log his nonce so we can use it to query bond blance
-        const DGOVNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
+        const DGOVNonces = (await bondContract.getNoncesPerAddress(buyer, 4)).map(n => n.toNumber());
         console.log("nonce: " + DGOVNonces[0]);
         let dgovbondBalance = await bondContract.balanceOf(buyer, 4, DGOVNonces[0] );
         
@@ -183,51 +146,260 @@ contract('Bank', async (accounts: string[]) => {
         console.log("DBITbondBalance: " + DBITbondBalance.toString());
 
         //we should have 3000 usdc bond
-        expect( dgovbondBalance.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+        expect(DBITbondBalance.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
         
+        //we should have minimum interest
+        expect( parseFloat(web3.utils.fromWei(dgovbondBalance, "ether"))).to.greaterThan(90);
+
         //now we have check bond balances, we check balances in apm
 
+        
         const s = await apmContract.getReserves(dbitContract.address, dgovContract.address);
         console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
 
         let APMbalanceDGOV = await dgovContract.balanceOf(apmContract.address);
         console.log("apmbalanceusdc :" + APMbalanceDGOV.toString());
-        expect( APMbalanceDGOV.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
 
         let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
         console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+        expect( APMbalanceDBIT.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+        
     })
 
-    it('buy Bonds: stakeForDbitBondWithEth', async () => {
-        usdcContract = await USDC.deployed();
-        usdtContract = await USDT.deployed();
-        bankContract = await Bank.deployed();
-        dbitContract = await DBIT.deployed();
-        wethContract = await WETH.deployed();
-        apmContract = await APM.deployed();
-        bondContract = await DebondBondTest.deployed();
+    it('stakeFordgovBondWithElse', async () => {
 
+        //mint usdc to the buyer
+        await usdcContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
+
+        //log his usdc balance
+        let balance = await usdcContract.balanceOf(buyer);
+        console.log("usdcBalance" + balance.toString());
+
+        //approve usdc and buy bonds
+        await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        let amount = await bankContract.stakeForDgovBondWithElse(USDC_FIX_6MTH_CLASS_ID, 4, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        
+       //log his nonce so we can use it to query bond blance
+        const dgovNonces = (await bondContract.getNoncesPerAddress(buyer, 4)).map(n => n.toNumber());
+        console.log("nonce: " + dgovNonces[0]);
+        let dgovbondBalance = await bondContract.balanceOf(buyer, 4, dgovNonces[0] );
+        
+        //same for usdc
+        const USDCNonces = (await bondContract.getNoncesPerAddress(buyer, USDC_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
+        console.log("nonce: " + USDCNonces[0]);
+        let UsdcbondBalance = await bondContract.balanceOf(buyer, USDC_FIX_6MTH_CLASS_ID, dgovNonces[0] );
+        
+        //query how much usdc buyer has now after buying
+        let USDCBALANCE = await usdcContract.balanceOf(buyer);
+        console.log("usdcBalance: " + USDCBALANCE.toString());
+
+        //log bond balances
+        console.log("dgovbondBalance: " + dgovbondBalance.toString());
+        console.log("USDCbondBalance: " + UsdcbondBalance.toString());
+
+        //we should have 3000 usdc bond
+        expect( UsdcbondBalance.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+        
+        //we should have minimum interest
+        expect( parseFloat(web3.utils.fromWei(dgovbondBalance, "ether"))).to.greaterThan(90);
+
+        //now we have check bond balances, we check balances in apm
 
         
-        await wethContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        const s = await apmContract.getReserves(usdcContract.address, dgovContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+        expect( APMbalanceUSDC.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+
+        let APMbalancedgov = await dgovContract.balanceOf(apmContract.address);
+        console.log("apmbalancedgov :" , APMbalancedgov.toString());
+        
+    })
+
+    it('buyforDbitBondWithElse', async () => {
+
+        //mint usdc to the buyer
+        await usdcContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
+
+        //log his usdc balance
+        let balance = await usdcContract.balanceOf(buyer);
+        console.log("usdcBalance" + balance.toString());
+
+        //approve usdc and buy bonds
         await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
-        await bankContract.stakeForDbitBondWithEth(10, DBIT_FIX_6MTH_CLASS_ID, 0, 2000, buyer, {from: buyer, value: web3.utils.toWei('2', 'ether')});
-
-        const DBITNoncesTest = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID));
-        console.log("nonce: " + DBITNoncesTest);
-
+        await bankContract.buyforDbitBondWithElse(1, 0, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        
+       //log his nonce so we can use it to query bond blance
         const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
         console.log("nonce: " + DBITNonces[0]);
         let dbitbondBalance = await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0] );
         
-        
+        //same for usdc
+        const USDCNonces = (await bondContract.getNoncesPerAddress(buyer, USDC_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
+        console.log("nonce: " + USDCNonces[0]);
         let UsdcbondBalance = await bondContract.balanceOf(buyer, USDC_FIX_6MTH_CLASS_ID, DBITNonces[0] );
         
+        //query how much usdc buyer has now after buying
         let USDCBALANCE = await usdcContract.balanceOf(buyer);
+        console.log("usdcBalance: " + USDCBALANCE.toString());
+
+        //log bond balances
         console.log("DBITbondBalance: " + dbitbondBalance.toString());
         console.log("USDCbondBalance: " + UsdcbondBalance.toString());
-        //expect( UsdcbondBalance.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+
+        //we should have 0 usdc bond
+        expect( UsdcbondBalance.toString()).to.equal(web3.utils.toWei('0', 'ether').toString());
+        
+        //we should have a minimum bonds
+        expect( parseFloat(web3.utils.fromWei(dbitbondBalance, "ether"))).to.greaterThan(3000);
+
+        //now we have check bond balances, we check balances in apm
+
+        
+        const s = await apmContract.getReserves(usdcContract.address, dbitContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+        expect( APMbalanceUSDC.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+
+        let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
+        console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+        
+    })
+    
+    it('buyForDgovBondWithDbit', async () => {
+
+        await dbitContract.setBankAddress(accounts[1]);
+        //mint dbit to the buyer
+        await dbitContract.mintCollateralisedSupply(buyer, web3.utils.toWei('100000', 'ether'), {from: accounts[1]});
+        await dbitContract.setBankAddress(bankContract.address);
+
+        //log his dbit balance
+        let balance = await dbitContract.balanceOf(buyer);
+        console.log("dbitBalance" + balance.toString());
+
+        //approve dbit and buy bonds
+        await dbitContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        await bankContract.buyForDgovBondWithDbit(0, 4, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        
+       //log his nonce so we can use it to query bond blance
+        const DGOVNonces = (await bondContract.getNoncesPerAddress(buyer, 4));
+        console.log("nonce: " + DGOVNonces[0]);
+        let dgovbondBalance = await bondContract.balanceOf(buyer, 4, DGOVNonces[0] );
+        
+        //same for DBIT
+        //const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, 0)).map(n => n.toNumber());
+        //console.log("nonce: " + DBITNonces[0]);
+        //let DBITbondBalance = await bondContract.balanceOf(buyer, 0, DBITNonces[0] );
+        
+        //query how much dbit buyer has now after buying
+        let DBITBALANCE = await dbitContract.balanceOf(buyer);
+        console.log("dbitBalance: " + DBITBALANCE.toString());
+
+        //log bond balances
+        console.log("dgovbondBalance: " + dgovbondBalance.toString());
+        //console.log("DBITbondBalance: " + DBITbondBalance.toString());
+
+        //we should have 0 usdc bond
+        //expect(DBITbondBalance.toString()).to.equal(web3.utils.toWei('0', 'ether').toString());
+
+        //we should have at least that amount of dgov
+        expect( parseFloat(web3.utils.fromWei(dgovbondBalance, "ether"))).to.greaterThan(3000);
+
+        
+        //now we have check bond balances, we check balances in apm
+
+        
+        const s = await apmContract.getReserves(dbitContract.address, dgovContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceDGOV = await dgovContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceDGOV.toString());
+
+        let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
+        console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+        expect( APMbalanceDBIT.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+        
+    })
+
+    it('buyForDgovBondWithElse', async () => {
+
+        //mint usdc to the buyer
+        await usdcContract.mint(buyer, web3.utils.toWei('100000', 'ether'));
+
+        //log his usdc balance
+        let balance = await usdcContract.balanceOf(buyer);
+        console.log("usdcBalance" + balance.toString());
+
+        //approve usdc and buy bonds
+        await usdcContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        let amount = await bankContract.buyForDgovBondWithElse(USDC_FIX_6MTH_CLASS_ID, 4, web3.utils.toWei('3000', 'ether'), 0, 2000, buyer, {from: buyer});
+        
+       //log his nonce so we can use it to query bond blance
+        const dgovNonces = (await bondContract.getNoncesPerAddress(buyer, 4)).map(n => n.toNumber());
+        console.log("nonce: " + dgovNonces[0]);
+        let dgovbondBalance = await bondContract.balanceOf(buyer, 4, dgovNonces[0] );
+        
+        //same for usdc
+        const USDCNonces = (await bondContract.getNoncesPerAddress(buyer, USDC_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
+        console.log("nonce: " + USDCNonces[0]);
+        let UsdcbondBalance = await bondContract.balanceOf(buyer, USDC_FIX_6MTH_CLASS_ID, dgovNonces[0] );
+        
+        //query how much usdc buyer has now after buying
+        let USDCBALANCE = await usdcContract.balanceOf(buyer);
         console.log("usdcBalance: " + USDCBALANCE.toString());
+
+        //log bond balances
+        console.log("dgovbondBalance: " + dgovbondBalance.toString());
+        console.log("USDCbondBalance: " + UsdcbondBalance.toString());
+
+        //we should have 0 usdc bond
+        expect( UsdcbondBalance.toString()).to.equal(web3.utils.toWei('0', 'ether').toString());
+
+        //we should have at least that amount of dgov
+        expect( parseFloat(web3.utils.fromWei(dgovbondBalance, "ether"))).to.greaterThan(3000);
+
+        
+        //now we have check bond balances, we check balances in apm
+
+        
+        const s = await apmContract.getReserves(usdcContract.address, dgovContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+        expect( APMbalanceUSDC.toString()).to.equal(web3.utils.toWei('3000', 'ether').toString());
+
+        let APMbalancedgov = await dgovContract.balanceOf(apmContract.address);
+        console.log("apmbalancedgov :" , APMbalancedgov.toString());
+        
+    })
+   
+
+    it('stakeForDbitBondWithEth', async () => {
+        
+        await wethContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        await bankContract.stakeForDbitBondWithEth(10, DBIT_FIX_6MTH_CLASS_ID, 0, 2000, buyer, {from: buyer, value: web3.utils.toWei('2', 'ether')});
+
+        const ETHNonce = (await bondContract.getNoncesPerAddress(buyer, 10));
+        console.log("nonce: " + ETHNonce);
+
+        const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
+        console.log("nonce: " + DBITNonces[0]);
+
+        let DBITbondBalance = await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0] );
+        let ETHbondBalance = await bondContract.balanceOf(buyer, 10, ETHNonce[0] );
+        
+        
+        console.log("DgovbondBalance: " + DBITbondBalance.toString());
+        console.log("ETHbondBalance: " + ETHbondBalance.toString());
+
+        expect( ETHbondBalance.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
+        expect( parseFloat(web3.utils.fromWei(DBITbondBalance, "ether"))).to.greaterThan(0.06);
+
 
         //now we have check bond balances, we check balances in apm
 
@@ -243,20 +415,128 @@ contract('Bank', async (accounts: string[]) => {
         const WETHbalanceOfAPM = (await wethContract.balanceOf(apmContract.address)).toString()
         console.log("WETH Balance of APM: " + WETHbalanceOfAPM);
 
+        expect( WETHbalanceOfAPM.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
     })
 
+    it('stakeForDgovBondWithEth', async () => {
+        
+        await wethContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        await bankContract.stakeForDgovBondWithEth(10, 4, 0, 2000, buyer, {from: buyer, value: web3.utils.toWei('2', 'ether')});
+
+        const ETHNonce = (await bondContract.getNoncesPerAddress(buyer, 10));
+        console.log("nonce: " + ETHNonce);
+
+        const DGOVNonces = (await bondContract.getNoncesPerAddress(buyer, 4)).map(n => n.toNumber());
+        console.log("nonce: " + DGOVNonces[0]);
+
+        let DgovbondBalance = await bondContract.balanceOf(buyer, 4, DGOVNonces[0] );
+        let ETHbondBalance = await bondContract.balanceOf(buyer, 10, ETHNonce[0] );
+        
+        
+        console.log("DgovbondBalance: " + DgovbondBalance.toString());
+        console.log("ETHbondBalance: " + ETHbondBalance.toString());
+
+        expect( ETHbondBalance.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
+        expect( parseFloat(web3.utils.fromWei(ETHbondBalance, "ether"))).to.greaterThan(0.06);
 
 
-    it('redeem Bonds too early should send error', async () => {
+        //now we have check bond balances, we check balances in apm
+
+        const s = await apmContract.getReserves(usdcContract.address, dbitContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+
+        let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
+        console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+
+        const WETHbalanceOfAPM = (await wethContract.balanceOf(apmContract.address)).toString()
+        console.log("WETH Balance of APM: " + WETHbalanceOfAPM);
+
+        expect( WETHbalanceOfAPM.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
+    })
+
+    it('buyforDbitBondWithEth', async () => {
+        
+        await wethContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        await bankContract.buyforDbitBondWithEth(10, DBIT_FIX_6MTH_CLASS_ID, 0, 2000, buyer, {from: buyer, value: web3.utils.toWei('2', 'ether')});
+
+        const ETHNonce = (await bondContract.getNoncesPerAddress(buyer, 10));
+        console.log("nonce: " + ETHNonce);
 
         const DBITNonces = (await bondContract.getNoncesPerAddress(buyer, DBIT_FIX_6MTH_CLASS_ID)).map(n => n.toNumber());
         console.log("nonce: " + DBITNonces[0]);
-        try {
-            await bankContract.redeemBonds(DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0], web3.utils.toWei('1000', 'ether'), {from: buyer});
-        } catch (e: any) {
-            assert.isTrue(true)
-        }
 
-        console.log("balance Bond D/BIT: AFTER " + (await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0])));
+        let DBITbondBalance = await bondContract.balanceOf(buyer, DBIT_FIX_6MTH_CLASS_ID, DBITNonces[0] );
+        //let ETHbondBalance = await bondContract.balanceOf(buyer, 10, ETHNonce[0] );
+        
+        
+        console.log("DgovbondBalance: " + DBITbondBalance.toString());
+        //console.log("ETHbondBalance: " + ETHbondBalance.toString());
+
+        //expect( ETHbondBalance.toString()).to.equal(web3.utils.toWei('0', 'ether').toString());
+        expect( parseFloat(web3.utils.fromWei(DBITbondBalance, "ether"))).to.greaterThan(2);
+
+
+        //now we have check bond balances, we check balances in apm
+
+        const s = await apmContract.getReserves(usdcContract.address, dbitContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+
+        let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
+        console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+
+        const WETHbalanceOfAPM = (await wethContract.balanceOf(apmContract.address)).toString()
+        console.log("WETH Balance of APM: " + WETHbalanceOfAPM);
+
+        expect( WETHbalanceOfAPM.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
     })
+
+    it('buyforDgovBondWithEth', async () => {
+        
+        await wethContract.approve(bankContract.address, web3.utils.toWei('100000', 'ether'), {from: buyer});
+        await bankContract.buyforDgovBondWithEth(10, 4, 0, 2000, buyer, {from: buyer, value: web3.utils.toWei('2', 'ether')});
+
+        //const ETHNonce = (await bondContract.getNoncesPerAddress(buyer, 10));
+        //console.log("nonce: " + ETHNonce);
+
+        const DGOVNonces = (await bondContract.getNoncesPerAddress(buyer, 4)).map(n => n.toNumber());
+        console.log("nonce: " + DGOVNonces[0]);
+
+        let DgovbondBalance = await bondContract.balanceOf(buyer, 4, DGOVNonces[0] );
+        //let ETHbondBalance = await bondContract.balanceOf(buyer, 10, ETHNonce[0] );
+        
+        
+        console.log("DgovbondBalance: " + DgovbondBalance.toString());
+        //console.log("ETHbondBalance: " + ETHbondBalance.toString());
+
+
+
+        expect( parseFloat(web3.utils.fromWei(DgovbondBalance, "ether"))).to.greaterThan(2);
+        
+
+
+        //now we have check bond balances, we check balances in apm
+
+        const s = await apmContract.getReserves(usdcContract.address, dbitContract.address);
+        console.log("here we print r0 after stakebonds : " + s[0].toString(), "here we print r1 after stake bonds :" + s[1].toString());
+
+        let APMbalanceUSDC = await usdcContract.balanceOf(apmContract.address);
+        console.log("apmbalanceusdc :" + APMbalanceUSDC.toString());
+
+        let APMbalanceDBIT = await dbitContract.balanceOf(apmContract.address);
+        console.log("apmbalanceDBIT :" , APMbalanceDBIT.toString());
+
+        const WETHbalanceOfAPM = (await wethContract.balanceOf(apmContract.address)).toString()
+        console.log("WETH Balance of APM: " + WETHbalanceOfAPM);
+
+        expect( WETHbalanceOfAPM.toString()).to.equal(web3.utils.toWei('2', 'ether').toString());
+    })
+
+
+    
 });
